Move signatureToVRS to utils so we can import it in the hub, change the type of "v" in vrs to number
